pyproject.toml
[project]
name = "prompt-opt-demo"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "deepagents",
    "langchain>=1.0,<2.0",
    "langchain-core>=1.0,<2.0",
    "langchain-anthropic",
    "langchain-openai",
    "langchain-chroma",
    "langchain-tavily",
    "langchain-text-splitters",
    "langsmith>=0.3.0",
    "python-dotenv",
    "pandas>=3.0.1",
]
